#4 fix: 수정사항 반영

+++ beilsang/beilsang/Model/HomeNetwork/ChallengeJoinTwoModel.swift
@@ -23,5 +23,5 @@ struct ChallengeJoinTwoData : Codable {
     let challengeId : Int
     let title : String
     let imageUrl : String?
-    let achieveRate : Int
+    let achieveRate : Float
 }
